asyncIterable: locate async iterator errors to the collection (#3899)
Currently, they are presented as errors in the next pending item.

Of course, there may or not be a next pending item; there may not be a
way to definitively know the reason for the error, such that the error
is a function of a problem with the collection itself.

As discussed in a prior incremental delivery WG meeting, this must be
changed for both the non-streaming and streaming cases; this PR makes
the minor adjustments addressing both cases.