Forbid @skip and @include directives in subscription root selection (#3974)
This is an implementation of
https://github.com/graphql/graphql-spec/pull/860

The spec calls for a separate `CollectSubscriptionFields` algorithm due
to executing without `variableValues`; but I figured that maintenance
would be easier with the algorithms synchronized.

Please feel free to make any changes you need to this PR; it's just to
get the ball rolling.

Related GraphQL WG action item (from 2021!):
https://github.com/graphql/graphql-wg/issues/695